move the parser back
