+++ parser/parser.go
@@ -1,4 +1,4 @@
-package model
+package parser
 
 import (
 	"bufio"
